requirements.txt
# Add your Python dependencies here, one per line
# Example:
# fastapi==0.95.2
# uvicorn[standard]==0.22.0

fastapi
uvicorn[standard]

# Google GenAI client
google-genai

# dotenv loader for config.py
python-dotenv

# Additional discovered dependencies
pydantic
pinecone

# Supabase client for database operations
supabase
